backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Redis
redis[hiredis]==5.2.1

# Auth & Security
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
passlib==1.7.4

# HTTP client (AI proxy)
httpx==0.28.1

# SSE streaming
sse-starlette==2.2.1

# Background tasks
celery[redis]==5.4.0

# Google OAuth
google-auth==2.37.0

# Utilities
python-multipart==0.0.20
email-validator==2.2.0
